requirements.txt
fastapi==0.116.1
uvicorn==0.20.0
pytesseract==0.3.13
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
transformers==4.24.0
torch==2.7.1
scipy==1.15.3
scikit-learn==1.1.2
pyspellchecker==0.8.3
fuzzywuzzy==0.18.0
pydantic==2.11.7
